requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
attrs==25.1.0
backoff==2.2.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
opencage==2.3.0
propcache==0.2.1
python-dotenv==1.0.0
requests==2.32.3
SQLAlchemy==2.0.23
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.18.3
flask-cors==4.0.0
flask-bootstrap==3.3.7.1
flask-caching==2.1.0
gunicorn==21.2.0
